refactor: export import
- Added StudyExportService for exporting study data to JSON format.
- Introduced StudyImportService for importing study data from JSON format.
- Created ExportImportRegistry to manage mappings between IDs and handles during export/import.
- Developed HandleGenerator and HandlePrefixes for generating consistent identifiers.
- Refactored StudyProtocolSerializer to utilize the new export/import services, simplifying the codebase.
- Ensured compatibility with existing data structures and serialization logic.
