+++ designer_v2/lib/domain/serialization/utils/handle_prefixes.dart
@@ -1,0 +1,13 @@
+class HandlePrefixes {
+  HandlePrefixes._();
+
+  static const String questionnaire = 'screening';
+  static const String question = 'screening_q';
+  static const String intervention = 'intervention_';
+  static const String observation = 'observation_';
+  static const String criterion = 'criterion_';
+  static const String consent = 'consent_';
+  static const String task = 'task_';
+  static const String period = 'period_';
+  static const String choice = '_opt';
+}
